LayoutTests/fast/mediastream/RTCPeerConnection-expected.txt

Tests the RTCPeerConnection constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new webkitRTCPeerConnection(null); threw exception TypeError: RTCPeerConnection argument must be a valid Dictionary.
PASS new webkitRTCPeerConnection(undefined); threw exception TypeError: RTCPeerConnection argument must be a valid Dictionary.
PASS new webkitRTCPeerConnection(); threw exception TypeError: Not enough arguments.
PASS new webkitRTCPeerConnection(''); threw exception TypeError: RTCPeerConnection argument must be a valid Dictionary.
PASS new webkitRTCPeerConnection({}); threw exception TypeError: Invalid RTCPeerConnection constructor arguments.
PASS new webkitRTCPeerConnection({iceServers:[]}); threw exception TypeError: Error creating RTCPeerConnection.
PASS new webkitRTCPeerConnection({iceServers:[{url:'stun:foo.com'}]}); threw exception TypeError: Error creating RTCPeerConnection.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'http:foo.com'}]}); threw exception TypeError: Error creating RTCPeerConnection.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:['stun:foo.com', 'turn:foo.com']}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}, {urls:['stun:foo.com', 'turn:bar.com']}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'none'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'relay'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'all'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], requestIdentity:'yes'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], requestIdentity:'no'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'foo', requestIdentity:'bar'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'none', requestIdentity:'bar'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'foo', requestIdentity:'no'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'none', requestIdentity:'yes'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'none', requestIdentity:'no'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'none', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'relay', requestIdentity:'yes'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'relay', requestIdentity:'no'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'relay', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'all', requestIdentity:'yes'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'all', requestIdentity:'no'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', credential:'x'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', credential:'x'},{urls:'stun:bar.com'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com', username:'user'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:'stun:bar.com', username:'user2', credential:'x'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}]}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', credential:'x'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', credential:'x'},{urls:'stun:bar.com'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com', username:'user'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:'stun:bar.com', username:'user2', credential:'x'}], iceTransports:'all', requestIdentity:'ifconfigured'}); did not throw exception.
PASS new webkitRTCPeerConnection({fooServers:[]}); threw exception TypeError: Invalid RTCPeerConnection constructor arguments.
PASS new webkitRTCPeerConnection({iceServers:true}); threw exception TypeError: Invalid RTCPeerConnection constructor arguments.
PASS new webkitRTCPeerConnection({iceServers:[1, 2, 3]}); threw exception TypeError: Invalid RTCPeerConnection constructor arguments.
PASS new webkitRTCPeerConnection({iceServers:[{}]}); threw exception TypeError: Error creating RTCPeerConnection.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'foo'}]}); threw exception TypeError: Error creating RTCPeerConnection.
PASS successfullyParsed is true

TEST COMPLETE
